org changes and style fix
